backend/requirements.txt
django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.7.0
pillow==11.1.0
face-recognition==1.3.0
numpy==2.2.3
opencv-python-headless==4.11.0.86
dlib==19.24.6 